Add DaysIntervalCalculator service and tests
Implemented a new service, DaysIntervalCalculator, for calculating the number of days between two dates. This service also validates the date order, throwing a LogicException if the start date is later than the end date. Corresponding unit tests have been added to ensure the correctness of calculations and exception handling. The DailyOccupancyRateCalculatorTest has also been moved to the Services directory.