Replace Carbon with CarbonImmutable in various files
In this commit, Carbon was replaced by CarbonImmutable in various files. This was done to address issues related to the mutability of Carbon instances. CarbonImmutable instances do not face the same issues as they are immutable and their state cannot be changed after creation. Besides Carbon replacement, minor adjustments were made to clean up and streamline code.